Move painter into symbology folder
